clean up styles and prevent pairing multiple tokentypes
