+++ libs/tup-components/src/accounts/ManageAccount.module.css
@@ -1,13 +1,5 @@
-.account-icon {
-    font-size: 32px;
-    margin-bottom: -5px;
-    margin-right: 0.5rem;
-}
-
-.account-section {
+.account-layout{
     padding-right: 30px;
-    display: flex;
-    flex-direction: column;
 }
 
 .account-header {
@@ -15,29 +7,30 @@
     border-bottom: 1px solid #707070;
     padding: 10px;
 }
+.account-header > i {
+    font-size: 32px;
+    margin-bottom: -6px;
+    margin-right: 0.5rem;
+
+}
 
 .account-body {
     padding-top: 10px;
     display: flex;
-    flex: 1;
     flex-direction: row;
     gap: 3%;
 }
-
-.tap-column {
-    flex: 1.5;
+.account-body > section {
     display: flex;
     flex-direction: column;
     gap: 20px;
     height: 100%;
 }
-
-.mfa-column {
+.account-body > section:nth-of-type(1) {
+    flex: 1.5;
+}
+.account-body > section:nth-of-type(2) {
     flex: 1;
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    height: 100%
 }
 
 .tap-header {
@@ -54,4 +47,15 @@
 
 .tap-button {
     width: 200px;
+}
+
+.tap-separator {
+    border-right: 1px solid #707070;
+    margin-top: 10px;
+}
+
+.mfa-options {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
 }
